added splitter

+++ src/bit_pack.rs
@@ -1,5 +1,5 @@
 use crate::bit::Bit;
 
-pub trait BitPack<const N: usize> {
+pub trait BitPack<const N: u8> {
     fn msb(&self) -> Bit;
 }
